filters makes api call to filtered announcements by page
