+++ src/pages/Academics/AcademicsDashboard.tsx
@@ -1,37 +1,16 @@
-import React, { useContext, useState, useEffect } from 'react';
+import React from 'react';
 import { faGraduationCap } from '@fortawesome/pro-light-svg-icons';
-import { UserContext } from '../../App';
-import { useAnnouncements } from '../../api/announcements';
-import { hasAudience } from '../../api/user';
 import Courses from '../../features/Courses';
 import AnnouncementContainer from '../../ui/AnnouncementContainer';
 import { MainGridWrapper, MainGrid, MainGridCol, SecondGridWrapper } from '../../ui/PageGrid';
 import PlannerItems from '../../features/PlannerItems';
 import AcademicCalendar from '../../features/AcademicCalendar';
-import PageTitle, { Title } from '../../ui/PageTitle';
+import PageTitle from '../../ui/PageTitle';
 import AcademicOverview from '../../features/AcademicOverview';
 import ResourcesCard from '../../features/ResourcesCard';
 import { AcademicSubNav } from './AcademicsSubNav';
 
 const AcademicsDashboard = () => {
-
-  const announcementType = 'academic'
-  const [events, setEvents] = useState<any>([]);
-  const user = useContext<any>(UserContext);
-  const announcements = useAnnouncements(announcementType);
-  
-  /* eslint-disable react-hooks/exhaustive-deps */
-  // Fetch data on load
-  useEffect(() => {
-    if (!user.loading && !announcements.loading) {
-      const announcementsToUse = announcements.data.filter(announcement =>
-        hasAudience(user.data, announcement)
-      );
-      setEvents(announcementsToUse);
-    }
-  }, [announcements.data, announcements.loading, user.data, user.loading]);
-  /* eslint-enable react-hooks/exhaustive-deps */
-
   return (
     <>
       <MainGridWrapper>
@@ -50,12 +29,11 @@ const AcademicsDashboard = () => {
         </MainGrid>
       </MainGridWrapper>
       <SecondGridWrapper>
-        {events.length > 0 && (
-          <>
-          <Title as="h2">Announcements</Title>
-          <AnnouncementContainer data-testid="academics-announcements" className="col-span-2" type={announcementType} events={events} />
-          </>
-        )}
+        <AnnouncementContainer
+          data-testid="academics-announcements"
+          className="col-span-2"
+          page="academics"
+        />
       </SecondGridWrapper>
     </>
   );
