+++ src/pages/Finances.tsx
@@ -1,9 +1,6 @@
-import React, { useContext, useState, useEffect } from 'react';
+import React from 'react';
 import { faDollarSign } from '@fortawesome/pro-light-svg-icons';
-import { UserContext } from '../App';
-import { useAnnouncements } from '../api/announcements';
-import { hasAudience } from '../api/user';
-import PageTitle, { Title } from '../ui/PageTitle';
+import PageTitle from '../ui/PageTitle';
 import ResourcesCard from '../features/ResourcesCard';
 import AnnouncementContainer from '../ui/AnnouncementContainer';
 import FinancialTransactions from '../features/FinancialTransactions';
@@ -11,26 +8,6 @@ import FinancialOverview from '../features/financial-overview/FinancialOverview'
 import { MainGridWrapper, MainGrid, MainGridCol, SecondGridWrapper } from '../ui/PageGrid';
 
 const Finances = () => {
-
-  const announcementType = 'financial'
-  const [events, setEvents] = useState<any>([]);
-  const user = useContext<any>(UserContext);
-  const announcements = useAnnouncements(announcementType);
-  
-  /* eslint-disable react-hooks/exhaustive-deps */
-  // Fetch data on load
-  useEffect(() => {
-    let announcementsToUse: any[] = [];
-
-    if (!user.loading && !announcements.loading) {
-      announcementsToUse = announcements.data.filter(announcement =>
-        hasAudience(user.data, announcement)
-      );
-    }
-    setEvents(announcementsToUse);
-  }, [announcements.data, announcements.loading, user.data, user.loading]);
-  /* eslint-enable react-hooks/exhaustive-deps */
-
   return (
     <div data-testid="finances-page">
       <MainGridWrapper>
@@ -46,12 +23,11 @@ const Finances = () => {
         </MainGrid>
       </MainGridWrapper>
       <SecondGridWrapper>
-      {events.length > 0 && (
-          <>
-          <Title as="h2">Announcements</Title>
-          <AnnouncementContainer data-testid="finances-announcements" className="col-span-2" type={announcementType} events={events} />
-          </>
-        )}
+        <AnnouncementContainer
+          data-testid="finances-announcements"
+          className="col-span-2"
+          page="finances"
+        />
       </SecondGridWrapper>
     </div>
   );
