+++ src/ui/AnnouncementContainer.tsx
@@ -1,7 +1,11 @@
-import React from 'react';
+import React, { useContext, useState, useEffect } from 'react';
 import styled from 'styled-components';
+import { useAnnouncements } from '../api/announcements';
 import EventCard from './EventCard';
+import { Title } from '../ui/PageTitle';
 import { breakpoints } from '../theme';
+import { UserContext } from '../App';
+import { hasAudience } from '../api/user';
 
 const AnnouncementContainerWrapper = styled.div`
   max-width: ${breakpoints[1024]};
@@ -16,19 +20,38 @@ const AnnouncementContainerWrapper = styled.div`
   }
 `;
 
-const AnnouncementContainer = ({ type, ...props }) => {
-  const events = props.events
+const AnnouncementContainer = ({ page, ...props }) => {
+  const [events, setEvents] = useState<any>([]);
+  const user = useContext<any>(UserContext);
+  const announcements = useAnnouncements(page);
 
-  if (!events.length) {
+  /* eslint-disable react-hooks/exhaustive-deps */
+  // Fetch data on load
+  useEffect(() => {
+    let announcementsToUse: any[] = [];
+
+    if (!user.loading && !announcements.loading) {
+      announcementsToUse = announcements.data.filter(announcement =>
+        hasAudience(user.data, announcement)
+      );
+    }
+    setEvents(announcementsToUse);
+  }, [announcements.data, announcements.loading, user.data, user.loading]);
+  /* eslint-enable react-hooks/exhaustive-deps */
+
+  if (events.length === 0) {
     return null;
   }
-  
+
   return (
-    <AnnouncementContainerWrapper {...props}>
-      {events.map(item => (
-        <EventCard key={item.id} itemContent={item} />
-      ))}
-    </AnnouncementContainerWrapper>
+    <>
+      <Title as="h2">Announcements</Title>
+      <AnnouncementContainerWrapper {...props}>
+        {events.map(item => (
+          <EventCard key={item.id} itemContent={item} />
+        ))}
+      </AnnouncementContainerWrapper>
+    </>
   );
 };
